+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -460,6 +460,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed ray traced point and spot lights shadows not rejecting istory when semi-transparent or colored.
 - Fixed a warning due to StaticLightingSky when reloading domain in some cases.
 - Fixed the MaxLightCount being displayed when the light volume debug menu is on ColorAndEdge.
+- Fixed issue with unclear naming of debug menu for decals.
 
 ### Changed
 - Color buffer pyramid is not allocated anymore if neither refraction nor distortion are enabled